Stability sweep round 2: CG NaN guard, Save race, schema leniency, Children IReadOnlyList
Four more parallel audit agents (2D utilities + persistence + the 1D
algorithm / domain / VM coverage gap from round 1) turned up six real
fixes after verifying each claim against the code.

1. ColumnGenerationSolver NaN guard (HIGH). The custom Simplex's Pivot
   already short-circuits on near-zero pivots, but cells touched by
   the near-zero divide can still leave NaN downstream. If a single
   dual cell goes NaN, `pi[i]` flows into the knapsack pricing, every
   `value > best` comparison silently fails because NaN > x is always
   false, and the CG loop exits with improved=false — making the LP
   look "converged" while we never priced anything. Now: detect NaN
   /Infinity in the duals immediately and break out of the CG loop so
   the LP-rounding fallback uses the columns we accumulated.

2. UserPreferences.Save tmp-filename race (HIGH). The previous "atomic"
   write used a fixed `path + ".tmp"` suffix, so two threads invoking
   Save concurrently (drag-drop + a UI scenario save firing in close
   succession) would both write the same .tmp and one would lose. Now
   the suffix is `.{ProcessId}.{Guid:N}.tmp` so each writer gets a
   unique temp file; the rename target is still the canonical path so
   the atomicity guarantee holds.

3. ScenarioService schema leniency (MEDIUM). Load1D / Load2D enforced
   exact-string equality on `cutting-stock-1d/v1`. Any future v2 bump
   would silently break every user's saved scenarios. Now they accept
   the `cutting-stock-1d/` prefix and only reject if the schema family
   is wrong (e.g. someone dropped a 2D file into the 1D Open dialog).

4. UserPreferences.Load error breadcrumb (LOW). Was a bare `catch
   { return defaults; }`. Now appends a single line to crash.log so a
   user reporting "my prefs keep resetting" leaves a forensic trail.

5. WasteLength widened to long (LOW). Matches the long TotalCost
   migration from earlier and removes one more remote overflow path.
   Touches benchmarks (LastWasteLength) and the AlgorithmComparison
   tuple type.

6. GuillotineNode.Children → IReadOnlyList (LOW). The List was init-
   only but Add()/Insert() etc. were still publicly callable, so any
   consumer could splice a tree post-construction. Now the property
   is IReadOnlyList<GuillotineNode> with Array.Empty as the default;
   PatternBuilder and GuillotineValidatorEdgeTests rewritten to use
   `new GuillotineNode[] { ... }` for the four sites that built trees
   via collection-initializer syntax.

False positives the audit flagged but verification cleared:
- GuillotineKnapsackDp |X|·|Y| memo OOM and GuillotineValidator
  recursive stack overflow — pathological inputs only, not seen in
  realistic loads. Defer until a real report.
- HasOverlap kerf negative — SolverOptions2D.Kerf setter rejects
  negatives at the boundary already.
- WithinSheet trim > sheet/2 — produces a degenerate sheet area, the
  bounds check still rejects all placements correctly.
- PatternPool FNV-1a collision — 64-bit hash, birthday paradox kicks
  in at 2^32 columns. Pool sizes never reach that.
- BuildDpItems eps — `1e-6` is well below any meaningful dual.
- Cross-platform LocalApplicationData / path case sensitivity — app
  is Windows-only (UseWPF=true), the Core path code is unused on
  non-Windows in practice.

Tests stayed green: 624 Core + 40 UI = 664/664. The
Save_LeavesNoTempFileBehind test was extended to match the new
unique-suffix pattern (`Directory.GetFiles(dir, baseName + ".*.tmp")`
must be empty).

CHANGELOG.md updated.

+++ CuttingStock.Core/Algorithms/Utilities/SolverUtils.cs
@@ -93,7 +93,7 @@ namespace CuttingStock.Core.Algorithms.Utilities
 
             result.WasteLength = result.CuttingPlans
                 .Where(p => p.Leftover < options.Gamma)
-                .Sum(p => p.Leftover);
+                .Sum(p => (long)p.Leftover);
 
             var weldGroups = result.CuttingPlans
                 .SelectMany(p => p.Cuts)
